Fixed implicit narrowing warning from code scan.

+++ maverick-synergy-common/src/main/java/com/sshtools/synergy/ssh/ChannelDataWindow.java
@@ -32,11 +32,11 @@ public class ChannelDataWindow {
 		this.windowSpace = initialWindowSpace;
 	}
 	
-	public synchronized void consume(long count) {
+	public synchronized void consume(int count) {
 		windowSpace -= count;
 	}
 	
-	public synchronized void adjust(long count) {
+	public synchronized void adjust(int count) {
 		windowSpace += count;
 		notifyAll();
 	}
